Adds estimation of the 370 um fluorescence value to wzgrowth
The threshold is no longer arbitrarily half the maximum in the
first section, but is now estimated based on estimate of where
the proximal channel boundary is, and the taking the fluorescence
value at that boundary as the reference value moving forward.

This approach yields the cleanest growth data so far as so far as
reducing noise.