Adds code to handle 'hot-spike' outliers in wzgrowth
Sometimes at the start due to NP aggregation you get a hot spike due
to the laser hitting a fluorescent blob exactly. We want to ignore those.

The code calculates the halfway point between the highest and lowest z indexes
where the PMT value exceeds the threshold. It then requires at least 5% of the
lowest values is below  the halfway value. Otherwise it considers the low
extreme values outliers and removes them one by one until the criterion is
satisfied.